% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_options.R
\name{add_p_tbl_summary_test}
\alias{add_p_tbl_summary_test}
\title{Argument option for tbl_summary %>%  add_p()}
\usage{
add_p_tbl_summary_test()
}
\value{
character vector
}
\description{
Argument option for tbl_summary %>%  add_p()
}
