% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_icon_ui.R
\name{add_popover_help}
\alias{add_popover_help}
\title{Add embed help link to input object}
\usage{
add_popover_help(input_object, link_text, link_url)
}
\arguments{
\item{input_object}{object to add popover help.}

\item{link_text}{text to show in popover help.}

\item{link_url}{url for desired destination.}

\item{id}{id for shiny module.}
}
\description{
Add embed help link to input object
}
