% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{generate_testdata}
\alias{generate_testdata}
\title{Testdata Creation with missing values}
\usage{
generate_testdata(n, with_date = TRUE, with_na = TRUE)
}
\value{
tibble return tibble data.frame
}
\description{
Testdata Creation with missing values
}
