% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_regression.R
\name{get_terms}
\alias{get_terms}
\title{Function for capture terms in model formula.}
\usage{
get_terms(mod)
}
\description{
This function takes a regression model object and return terms of model formula. This function intended to use in the interactive tbl_regression add-in for include argument.
}
