% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_label.R
\name{labelModifyDropDownServer}
\alias{labelModifyDropDownServer}
\title{Module: Server for labelling dropdown menu}
\usage{
labelModifyDropDownServer(id, label_data)
}
\description{
Module: Server for labelling dropdown menu
}
