% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_load_variable.R
\name{load_environmental_variable}
\alias{load_environmental_variable}
\title{Internal function for load variable from environment.}
\usage{
load_environmental_variable(type = "data.frame")
}
\arguments{
\item{type}{Either Currently "data.frame" (for tbl_summary) only.}
}
\description{
Internal function for load variable from environment.
}
