% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_regression_addin.R
\name{tbl_regression_addin}
\alias{tbl_regression_addin}
\title{Function to run tbl_regrerssion add-in.}
\usage{
tbl_regression_addin()
}
\description{
This function is to call shiny app for tbl_regression
}
