% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_summary_addin_ui.R
\name{tbl_summary_addin_dropdown_ui_add_column}
\alias{tbl_summary_addin_dropdown_ui_add_column}
\title{generate dropdown UI for add_p for tbl_summary_addin}
\usage{
tbl_summary_addin_dropdown_ui_add_column()
}
\value{
list
}
\description{
generate dropdown UI for add_p for tbl_summary_addin
}
