% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_summary_addin_ui.R
\name{tbl_summary_addin_dropdown_ui_theme}
\alias{tbl_summary_addin_dropdown_ui_theme}
\title{generate dropdown UI for theme setting for tbl_summary_addin}
\usage{
tbl_summary_addin_dropdown_ui_theme()
}
\description{
generate dropdown UI for theme setting for tbl_summary_addin
}
