% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_summary_addin_ui.R
\name{tbl_summary_addin_sidebar_ui}
\alias{tbl_summary_addin_sidebar_ui}
\title{generate sidebar UI for tbl_summary_addin}
\usage{
tbl_summary_addin_sidebar_ui()
}
\value{
list
}
\description{
generate sidebar UI for tbl_summary_addin
}
