% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_options.R
\name{tbl_summary_type}
\alias{tbl_summary_type}
\title{Argument option for tbl_summary:type}
\usage{
tbl_summary_type()
}
\value{
character vector
}
\description{
Argument option for tbl_summary:type
}
