% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_options.R
\name{test_argument_for_add_p_tbl_summary}
\alias{test_argument_for_add_p_tbl_summary}
\title{Argument option for tbl_summary %>%  add_p()}
\usage{
test_argument_for_add_p_tbl_summary()
}
\value{
character vector
}
\description{
Argument option for tbl_summary %>%  add_p()
}
