% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_load_variable.R
\name{variableLoaderModalServer}
\alias{variableLoaderModalServer}
\title{Server Module to select variable from target environment.}
\usage{
variableLoaderModalServer(id, target_type = "data.frame")
}
\arguments{
\item{id}{id for shiny module}

\item{target_type}{use "data.frame" for tbl_summary}
}
\description{
Server Module to select variable from target environment.
}
