% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_label.R
\name{get_terms}
\alias{get_terms}
\title{Function for obtain terms from model object}
\usage{
get_terms(mod)
}
\arguments{
\item{mod}{model object}
}
\description{
Function for obtain terms from model object
}
