% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_label.R
\name{labelModifyDropDownServer}
\alias{labelModifyDropDownServer}
\title{Module: Server for labelling dropdown menu}
\usage{
labelModifyDropDownServer(id, label_data)
}
\arguments{
\item{id}{an id}

\item{label_data}{label_data to display in the module.}
}
\description{
Module: Server for labelling dropdown menu
}
