% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_load_variable.R
\name{variableLoaderModalUI}
\alias{variableLoaderModalUI}
\title{UI Module to select variable from target environment.}
\usage{
variableLoaderModalUI(id)
}
\arguments{
\item{id}{id for shiny module.}
}
\description{
UI Module to select variable from target environment.
}
